Updated openstack/openstack Project: openstack/nova  07c93c897e60a6eb9487dc770522e1424794ee34
Replace metaclass registry with explicit opt-in registry from oslo

This patch replaces Nova's magical metaclass object registry with the
one from oslo_versionedobjects.

Most of the changes here are related to test objects which previously
(and annoyingly) were automatically registered. It takes a more explicit
action to register these now, but also means we don't have to worry
about the version hash tracking picking up such objects.

There is a change to nova.exception here to copy the readonly exception
from the library, which is purely a tactic to reduce the amount of stuff
we have to change in this patch. The next patch will remove that and do
the necessary test conversions.

Related to blueprint use-oslo-objects

Change-Id: If3465f995dd05d2c01f01469b3ad9166c8362e8c