Updated openstack/openstack Project: openstack-infra/shade  36ffb13f28b5ef9c778ea3e1c6fe0875e9459fa6
Switch to SwiftService for segmented uploads

Swift caps individual file sizes at various levels, so we need to adapt
to this and upload in segments. swiftclient includes the SwiftService
object to do just that.

This adds a functional test to try to upload files to swift.  Because
sending 5GB of nulls does take a while, we're just going to try a
segmented upload and an unsegmented one, and see how that goes.

Note that this test coverage found a missing ObjectDelete task which
has been added. We also add the ObjectCapabilities task so that we can
make sure segment_size is never higher than max_file_size.

Change-Id: I43903235576245a316ae1921505237e9298b5200