Updated openstack/openstack Project: openstack-infra/devstack-gate  1ed9a9f426b8f1f0cb3ae4119b52fbe20a276b25
Add configuration switches for features that make Neutron unstable

Both using pymysql and multiple api workers unveiled a number of races that
occur when accessing the DB layer. We should enable these features in an
unstable configuration for Neutron so that we can shake them out and go back
to sanity.

Related-bug: #1464612
Related-bug: #1432189

Change-Id: Ib6346d53a1a5d99d868ee3ded0ce112b9d68ab51