Updated openstack/openstack Project: openstack/nova  2ea9c61f68d1d6e8e3f8e031898bbe1b6d90a169
hypervisor support matrix: add status "unknown"

When a new feature gets added to the hypervisor support matrix it
was not possible to state that the hypervisor did not make a support
statement yet. You had to choose between "complete", "partial" and
"missing" wich could be missleading.
This change set adds the status "unknown" which could make the
statements more precise.

Change-Id: If94fc772796742de88af5893839945834b9d3eb6