Updated openstack/openstack Project: openstack/neutron  b341440f1ee4fac00c3050edf2314541e0e6e826
Fix Enum usage in 589f9237ca0e_cisco_n1kv_ml2_driver_tables

PostgreSQL is more sensitive with types than MySQL, it creates a
separate type when a Enum is created. In migration 589f9237ca0e
type profile_type is trying to be created, but the type with such
name was already created in havana_initial migration.

The solution for this is not to create type in 589f9237ca0e
migration when dialect is PostgreSQL and use already created.

Closes-bug: #1463301

Change-Id: I66e74d50cc70673de8635616076779cc20cde113