Updated openstack/openstack Project: openstack/horizon  c16ca0668abb76c5af7bb093245df4863249d9fb
Merge tag '2015.1.0'

This is a null-merge of the 2015.1.0 release tag back into the master
branch so that the 2015.1.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: Ifbe5418d35b042c5e6ef72443c21340a524db3fa