Updated openstack/openstack Project: openstack/horizon  f7d5a619811b72645dc4f0d3755b0713daa4cef2
Merge tag '2014.2'

This is a null-merge of the 2014.2 release tag back into the
master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: I0127fe09d211cec231dab26d5987dafe720c91dd