Updated openstack/openstack Project: openstack/neutron  1435e0ca1412afb1ac06fb242ee3fd9e655d1336
Reflect project moves from stackforge to openstack.

Several git repos were just moved from stackforge to openstack.
Reflect the move in various places where the URL was in docs and
comments.  In passing, also change URLs to git.openstack.org instead
of github, as that is the official home of all of these repos.

Change-Id: I6c79a192d6604cef01e88d5b305fcc2b0f9c6b30
Co-Authored-By: Kyle Mestery <mestery@mestery.com>
Signed-off-by: Russell Bryant <rbryant@redhat.com>
Signed-off-by: Kyle Mestery <mestery@mestery.com>