Updated openstack/openstack Project: openstack-infra/system-config  e50c0cc6f6caa2b3e582eec49b948b51e4ae4e30
use optimized url for zuul status

This returns the optimized url for zuul, which should dramatically
drop the amount of data transfered out of zuul. It also uses the ajax
method which lets us ensure that we don't schedule another pass
through the loop until we completed the current request.

Add support for jquery visibility optimization

Change-Id: I0e50f07fd9165e7a45c2712f6b1c4b3601759cba