Updated openstack/openstack Project: openstack/cinder  d873ceb70448ad0ab1d88a2b0bdff9cfccd54f30
Remove pretty_tox and use ostestr

This patch dumps pretty_tox.sh and gets Cinder on board
with using ostestr.  We're setting a concurrency value
of 6 (instances in the gate are 6 VCPU's), and things
seem to be stable at this value.

For folks running on 'smaller' system, it's still possible
to run in parallell; > ncpu's but if there are problems you
can still specify concurrency in your tox argument.  Tox will
take the last occurence of the conncurrency argument; so for
example if you specify "tox -epy27 -- --concurrency=4" the
interpretted value for concurrency in the run will be 4.

Closes-Bug: #1459852

Change-Id: Iae12b42e915da22c863781015512df4f30aa95ca