Updated openstack/openstack Project: openstack/tempest  7ebf503a9bd386335042da8a1a509966ed56c993
Add passenv option to tox.ini

Tox 2.0 introduced env variable separation, meaning that env vars set
in the env calling tox will not automatically be in the test env. This
breaks some assumptions we have in tempest. This commit adds all the
env vars which are used by testr or tempest directly for running tests
in the hope of making things work on tox 2.0 and older versions.

Change-Id: Icfe4bb5df0e48cbc9fcdbc602656672295d1887e