Updated openstack/openstack Project: openstack/tripleo-heat-templates  48d07db384a3e8a36a2a92e95ef3ad4ed5c6071e
Add BondInterfaceOvsOptions parameter

This patch adds a new BondInterfaceOvsOptions to the
bond-with-vlans network config templates. This can
be used to configure things like LACP or the bonding mode
via a nested stack heat parameter.

The patch also removes the hard coded ovs_options relating
to both bond-with-vlans and single-nic-with-vlans configurations
which do not actually require this setting to be hard coded
because by default OVS trunks vlan ports automatically.

Change-Id: I3effbccba8ed7ed28d6ba715e5709275d4e7f984