Updated openstack/openstack Project: openstack/nova-specs  ba11ff242b5dd7b1d18b6cc29f13568388ca2a5d
Sort instances if possible inside an host aggregate

Simple use case: Reduce the licence cost
Windows licences are billed by host or by socket. The goal is to stack all
Windows instances on the same hosts as much as possible without creating a
new host aggregate.

As much as possible means this is a best effort feature, if nova can place
this new Windows instance on a compute where other Windows are running, do
it. If not, spawn it where you want or on a empty compute to identify it
as « preferred widows compute » for future windows spawn.

This use case could be extended to many other situations, every time you
have a preferred situation with no hard restrictions.

Co-Authored-By: Michael Still <mikal@stillhq.com>

Change-Id: I261bcb802573406f0aca5a87606e4dcca877301b