Updated openstack/openstack Project: openstack/manila  cedb9f452791b0d70ffc7a9883e6ec483cd22652
Merge tag '2015.1.0'

This is a null-merge of the 2015.1.0 release tag back into the master
branch so that the 2015.1.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: I13616ea7dd533242c12b9bf4a8af9f45c3d1afa3