Updated openstack/openstack Project: openstack/tempest  98f62ee80673532782bd77158512d8f94d67c128
Made a several-seconds checking&waiting for object to be deleted

There is a bug https://bugs.launchpad.net/tempest/+bug/1452915
that describes failing test_get_object_after_expiry_time test
because in real clouds hardcoded 3 seconds waiting for the
container is not enough. Replaced hardcoded 3 seconds with the
checking and waiting in a loop for object to be deleted.

Change-Id: If90d9ef047344e2e094aa1b143c93c9fa4b56fbd
Closes-Bug: #1452915