Updated openstack/openstack Project: openstack/glance  ccf2922988b6d58e1c33a7157c012bb6b093da30
Use ConfigFixture to ensure config settings are reverted

Update the test configuration management code to use a ConfigFixture so
that configuration changes are rolled back at the end of each test.

Also fix the initialization order in TestSwiftParams to ensure that the
base class is configured before local customizations are applied.

Change-Id: I27620e3d4caf95e2c75fbf2cf2716ae1337030c7