Updated openstack/openstack Project: openstack/cinder  826cac77c401d8be9ab7569b94d398264ecc549c
Use correct rtslib namespace for newer versions

For rtslib-fb versions 2.1.52 and newer its own namespace (rtslib_fb) is
recommended instead of the older one (rtslib).

On versions with this new namespace a deprecation warning will be
thrown: "'rtslib' package name for rtslib-fb is deprecated, please
instead import 'rtslib_fb'"

This patch uses this new namespace when possible and falls back to the
older namespace if not.

Change-Id: I984105458da4d7b8736bf7c83919cbe5feb9c85a
Closes-Bug: #1460814