Updated openstack/openstack Project: openstack/python-cinderclient  77d1106dcd19434ec809a9296991413862c23b54
The is_public filter in VolumeTypeManager.list broke the find in
VolumeTypeAccessManager, which caused all commands which should have taken
a volume_type name to fail.  I have choosen to fix this by effectively
removing the filter in the client (and the --all argument to type-list).
This is OK, since Cinder implements the filtering by user anyway.  The
consequence of this change is that the Admin user will always see the
entire list (including private types) every time they execute type-list

Change-Id: I5c9a693711db2918abbd3bbfae53e48162a01c3c
Closes-Bug: 1462104