Updated openstack/openstack Project: openstack/nova  48f1f12a5b7d3f347c322977c2a6e289ae5854e6
Handle InstanceNotFound when sending instance update notification

When Tempest (or any automated client) is waiting for an instance to be
ACTIVE but goes to ERROR state, it immediately fails a test and deletes
the instance.  The instance could be deleted by the time the instance
update notification processing happens so there is a race where that
could fail with InstanceNotFound.  That will get traced to the conductor
logs and we should try to keep the conductor logs as clean as possible
from expected exceptions.

This handles InstanceNotFound and just logs it at debug rather than
trace out the exception.

Closes-Bug: #1455640

Change-Id: Iaee147051afb59f7991f8dfef3bd3072c28f829d