Updated openstack/openstack Project: openstack/tempest  c2230b62f028395a9abc30d734473e772487b4d0
Apply a naming rule of GET to compute clients(n*)

[GET /resources] methods should be "list_<resource name>s"
or "show_<resource name>", so this patch applies the rule
to compute clients which name is "networks".

In addition, this patch moves "name" filter of list_networks
to caller side because this filter is used at one place.

Partially implements blueprint consistent-service-method-names

Change-Id: Ic2f7b8a2724035f0b2cd9def13cd7f7fc1e9924d