Updated openstack/openstack Project: openstack/heat  5f9c0831988296a9cf72fc76171c78d509f987c1
Add _add_event at the end of update_task

This patch adds call of _add_event method at the end of the update_task.
So now in event-list output user can see, that updating stack is
completed or failed.
Also status reason for update_task were changed to be inline with status
reason in the main stack_task method.

Method set_state is not used due to using stack.store method in the end
of update to avoid duplicate update.

Change-Id: Ic94c95f1558fd67f351696645426c00de8db1bf8
Closes-Bug: #1465321