Updated openstack/openstack Project: openstack/ceilometer-specs  05101cae87655cdd8871a9a4b08e775d3466147e
Spec to remove the pipeline from the api server

As described in the spec, having a pipeline in the api server is a
burden on several dimensions. The spec proposes to replace the
pipeline, used only by a single API resource, with producing
notifications that would be received by the notification-agent,
which would use its own pipeline for transformations.

blueprint api-no-pipeline
Co-Authored-By: liusheng <liusheng@huawei.com>

Change-Id: I099c127eea5d747526800ae69c4abf48ceba92ea