Updated openstack/openstack Project: openstack-infra/yaml2ical  746a15bcde2aa98d7371c5fb90b81b8d5d52ae53
Add project and agenda URLs to the bottom of calendar descriptions

agenda_url is already used in the meeting file template, now it will
be attached to every calendar event.

The new key project_url is for the page describing the overall
project and it too, if present will be in the calendar event.

Closes: https://storyboard.openstack.org/#!/story/2000288

Change-Id: Ic9f823bf744b27f42359858e55557b36fcab8f0a
Signed-off-by: Brian (bex) Exelbierd <bex@pobox.com>