Updated openstack/openstack Project: openstack/sahara  c85830c44e4f1d2bba5eab4c3d67a22bcb9e3e52
Also install alembic_migration folder

In Debian, doing "python setup.py install" fails to install the
alembic migration folder. This patch, which we currently carry
in the debian/patches folder, fixes it. Please allow it to reach
upstream, as it simplifies a lot our package maintainer's life if
it is there.

Change-Id: I640aa91c9c917483e6656f550c191c4d58636070