Updated openstack/openstack Project: openstack-infra/project-config  6acfe867e09eee8b26184e8b327c2293392621c9
Run requirements update entrypoint from a venv

The requirements update.py script is growing its own dependencies so
we need to install it into a virtualenv and run it from there. Also,
go ahead and switch to invoking its entrypoint since it'll be
installed for us now. Also run the generate-constraints entrypoint
from a virtualenv similarly.

Change-Id: I16e967356d5c56f1474ee661b954b3db11a608cb