Updated openstack/openstack Project: openstack/trove  3d5df409cdac535feac8b71cd8d5331595ebcd9a
Remove nova_proxy_admin_user from trove guest

The nova_proxy_admin_user isn't required on the guest agent, and can
be removed. This is used to look up the admin context, but doing this
is unnecessary since the user context is what should be passed to the
conductor. For heartbeats, a context is not needed and just passing an
empty dict should suffice.

Change-Id: Ic41abe58fcf4b7022360b59dad0bc3af1205b9ba
Closes-bug: 1339349
Co-authored-by: Nikhil Manchanda <SlickNik@gmail.com>
Co-authored-by: Amrith Kumar <amrith@tesora.com>