Updated openstack/openstack Project: openstack-infra/project-config  ea6eb712c5c5d6895ab4777f4ddeb2fb13f21e25
Submit new project puppet-openstack-cookiecutter

During last summit, it was decided during the puppet-openstack design
session that the community will provide an easy way for new project to
bootstrap their puppet module with a 'compliant' design.

A cookiecutter template was created[1], after mentioning it on the
list[2] and during the weekly meeting[3] it is a go from the community that
this project should be on stackforge so everyone can contribute to it
and so it is on the scope it belongs.

[1] https://github.com/enovance/puppet-openstack-cookiecutter
[2]
http://lists.openstack.org/pipermail/openstack-dev/2015-May/064608.html
[3]
http://eavesdrop.openstack.org/meetings/puppet_openstack/2015/puppet_openstack.2015-06-02-15.00.html

Change-Id: I18abf0f35f2787b9252a02a9293a669ab1389319