Updated openstack/openstack Project: openstack-dev/devstack  ccd4c2e1775a214967ca287729facfc21bd4e59d
Install python-libvirt for ceilometer only if n-cpu enabled

Nova is responsible for installing the libvirt package (if it is being
used). It is required by python-libvirt but python-libvirt only required
in ceilometer if nova compute is being used. There are some usage
scenarios where nova compute is not being used so in that case don't
install python-libvirt.

Change-Id: I0db66f1c0526e24ade98de85989a5ed8d37f0c4f