Updated openstack/openstack Project: openstack/cinder  a2ae130019b1a6a1adf93f45109b3aa9f2acd716
LVM add multiattach flag capability

This patch adds reporting of the multiattach capability
in the lvm driver.   Currently the Cinder scheduler does filtering
on hosts at create time for the --allow-multiattach option being
passed into the create request from the client.   In order for someone
to be able to create an LVM volume that can do multiattach, the LVM
driver needs to report the capability.

This patch's purpose is to allow LVM volumes to be multiattachable
for the folks working on the Nova patches to get Nova to do multiattach.
Without this patch, they can't do end to end testing.

Change-Id: I91942569a12d7c7637c7d9ffb8bbff0daeaf6a97