Updated openstack/openstack Project: openstack/tripleo-heat-templates  7691bacbde2eac6b18800096c46dbb9141a64481
Remove NO_SIGNAL from puppet role templates

Currently we use NO_SIGNAL on both the NetworkConfig and subsequent config
deploying the data associated with the role.  This means there is a risk that
should the NetworkConfig do anything interruptive (os-net-config can do
interface renaming based on discovery data for example) the role configuration
config could fail, and we'd never know until some later error occurs.

Additionally, we need to be sure that the heiradata deployed by each of the role
specicific configs is actually in-place before proceeding with any of the cluster
configuration - atm this works due to the inherent delays involved deploying to
bare-metal, but there's still a theoretical race if very fast deployment backends
(I'm thinking containers, e.g lxc backend to nova or something) were used instead.

Essentially, we should never be using NO_SIGNAL unless we want to ignore failure,
which AFAICT is not the case in this instance.

Change-Id: I0dbbcc87fb8df8e6bc4775c39fa616b0d0713464