Updated openstack/openstack Project: openstack/nova  8e5c68c707ccd161efe691215062515b7095925b
test: add MatchType helper class as equivalent of mox.IsA

The mox test library had a helper method mox.IsA(SomeType) which
allowed unit tests to assert that a method call argument had a
particular type. The mock test library has a mock.ANY helper
which is often used, but nothing that allows a stricter checker
for a specific type. This patch introduces a MatchType class
which is similar to mock.ANY, but restricted to a single type.

So instead of using lots of mock.ANY parameters

   mock_some_method.assert_called_once_with(
                  "hello",
                  mock.ANY,
                  mock.ANY,
                  "world",
                  mock.ANY)

It becomes possible to be stricter

   mock_some_method.assert_called_once_with(
                  "hello",
                  MatchType(objects.Instance),
                  mock.ANY,
                  "world",
                  MatchType(objects.KeyPair))

Change-Id: I3a1ca33500ef8007b6c496e4e0917d7de07ac40a