Updated openstack/openstack Project: openstack/horizon  dc3d7e57f90c6dc4c746bcafef3fcbc1a8310548
Allow to attach FIP without a router

The Default Neutron implementation needs a router with a gateway to associate a
FIP. So by default a topology check will be performed by horizon to list only
VM ports attached to a network which is itself attached to a router with an
external gateway. This is to prevent from setting a FIP to a port which will
fail with an error.

Some Neutron vendors do not require it. Some can even attach a FIP to any port
(e.g.: OpenContrail) owned by a tenant.

This patch adds a setting allowing to bypass the topology check in order to let
the user associate a FIP to an instance on a subnet with no router if the
Neutron backend allows it.

Change-Id: I1bf8bd0113a154c5e0bf0f75065b911f15db3e83
Closes-Bug: #1462902