Updated openstack/openstack Project: openstack/heat  a9a4e68319a2d30ce9ad8dce6d0422c68cb24844
Reusable empty template

Refactored the empty template and added it to template module so that it
can be re-used else where. In convergence it is needed for rolling back
when create fails. Also needed for deleting stack, that is, update the
stack with empty template.

The empty_template method can be passed a template header (version) in
order to get an empty template of that version, otherwise by default, an
empty HOT template is returned.

Change-Id: Ic59bd2d4b9cfed1efd1973af0c4ec9274bcc204a