Updated openstack/openstack Project: openstack/keystone  6b274951828955f65d35eb2fb9092f15d15a7bd7
Fix tests failing on slower system

Many of these auth tests will fail on a slower system or when
using the debugger. If the test takes too long it fails when
token timestamps are compared.

The strategy to fix this is to take control of the clock and not
depend on the external environment.

Change-Id: Ib00ba4aebfe4886c3d93e77f7b216139bdf2acf8
Closes-Bug: 1464366