Updated openstack/openstack Project: openstack/taskflow  2fa4af7a24584b211ca549f2df715c2d126360c8
Split-off the additional retry states from the task states

Split the states that are not task states (but are retry
states) into there own additional set and then use that set
and a new function to validate the transition at other locations
in the code-base.

This makes the transitions that are valid for tasks/retries
easily viewable, more easy to read and understand, and
more correct (instead of being a mix of task + retry atom
transitions and states).

Change-Id: I9515c19daf59a21e581f51e757ece2050f348214