Updated openstack/openstack Project: openstack/neutron  3454c87034e13655f2ab5489530cf0437cda66f4
Add devices to update in RPC call security_groups_provider_updated

When a security_groups_provider_updated is received then a global
refresh of the firewall is performed. This can be avoided if the
plugins pass as parameter of the call the devices that belongs to
the network updated.

Partially-Implements: blueprint restructure-l2-agent
Change-Id: I1e78f3a5ec7e5c5bcba338a0097566422411ef7e