Updated openstack/openstack Project: openstack/heat  cc1804d397b4e70b299b977dc73a31c16f343498
Missing template file on update fails silently

When the template file for a template/facade/provider resource is
not found in the new files list when doing a stack update, Heat
silently falls back to using the existing template (i.e. so no
change will be made).
The cause is that the update code shares the same code path for
obtaining the template as the init code. The init code needs to
have a fallback, so that loading the (parent) stack never fails,
however in the update case it is undesirable.

Change-Id: Id12b63d9f04caba8e9325aed0ee2acf51ffe0c4c
Closes-Bug: #1452520