Updated openstack/openstack Project: openstack/tripleo-heat-templates  d29411849004603a98cd45aba843eabe538a18c0
Rename ServiceNetMap: NeutronLocalIp

This patch renames the NeutronLocalIp option to be called
NeutronTenantNetwork. This is more consistent with
all of the other ServiceNetMap settings which end in
'Network' and initial end user feedback found the
old name a bit cryptic as well.

This is the network for neutron tenant traffic so lets
just name it that.

Change-Id: Id49afe75c372887453413c092190a5775aa3e1ee