Updated openstack/openstack Project: openstack-infra/project-config  72e348e547cd5010ef32c54ed8b7d2eb963a889d
Create Magnum UI repo

The Magnum community has evaluated inclusion of this project and decided on a
governance approach for managing this project in Magnum.

Reference:
http://lists.openstack.org/pipermail/openstack-dev/2015-June/066701.html

The Technical Committee is evaluating the addition of magnum-ui to the openstack
git namespace in a Magnum governance change.

Reference: https://review.openstack.org/#/c/192804/

Change-Id: Ice2e14bf999f79ca0c0718948db616456faab907