Updated openstack/openstack Project: openstack/requirements  407ebb769da57dc98e10edb4385445fb46e425a4
Increase testtools version to support unittest2

The patch increase minimal testtools version from 0.9.36 to
minimal version where testtools.TestCase is inherited from
unittest2.TestCase (instead of unittest.TestCase).
It allows to rely on single implementation of TestCase(unittest2
only) and use some features of unittest2(for example,
some projects like Sahara, Heat use self.assertRegex in test cases
but this method is supported in unittest2 only. So we need to
increase the minimal version of testtools to support this method).

Change-Id: I24d13c7701b7314e3599c94701af4b38d2dabefb