Updated openstack/openstack Project: openstack-infra/project-config  2f26f3c45ff7f7c3378a6900fae346d4e9401172
Fix Manila jobs for running on stable/juno branch

Manila uses 'plugin' approach in Devstack since Kilo. So, when Manila Tempest
jobs are running against 'stable/juno' branch of Manila it is just not
installed and all Tempest tests are skipped because of that. So, make jobs
distinguish two different approaches for setup of Manila using Devstack.

Change-Id: Ifb020fe8693fd4f014e944aad0c96a1e67843e4e