Updated openstack/openstack Project: openstack/sahara-specs  2c6cf61aa80eeaf50be0c3b71952e40526e43694
[EDP] Allow editing job binaries

Currently there is no way to edit a job binary. If a path needs
to be changed, for example, the job binary must be deleted and a new
one created. The most common use case is a situation where a user
creates a job binary, runs a job, and receives an error from the job
because the path does not exist. Although it is not strictly necessary,
editable job binary objects would be a convenience when a user needs to
correct a path or credentials.

Change-Id: Ic376796e32e0164ada86626438e79c5d2bbf59e9
blueprint: edp-edit-job-binaries