Updated openstack/openstack Project: openstack/sahara-specs  02ece704cd48a473c843c13e5794e1cfbd613ae2
[EDP] Allow editing datasource objects

Currently there is no way to edit a datasource object. If a path needs
to be changed, for example, the datasource must be deleted and a new
one created. The most common use case is a situation where a user
creates a datasource, runs a job, and receives an error from the job
because the path does not exist. Although it is not strictly necessary,
editable datasource objects would be a convenience when a user needs to
correct a path or credentials.

blueprint edp-edit-data-sources

Change-Id: Idaef58f596b59a67f1fe953f9cd99a3a10932382