Updated openstack/openstack Project: openstack/neutron  f817faca2b669076b4c263928c0362995f675fd0
Fixed the only sphinx warning in docs

We made previous attempt to get rid of all warnings, but it turned out
that gate does not execute tox's docs job but runs build_sphinx
directly.

The latter behaviour should be fixed, but while at it, we
should prepare neutron job to be executed in gate by cleaning up all
warnings.

Closes-Bug: #1466554

Change-Id: I8c265eae2175425568479116d1faef7d87fdcc02