Updated openstack/openstack Project: openstack/ironic  8f966109abf8954cea365dda09e52c2e3c7573ab
Added documentation to Vagrantfile

A couple of comments during my latest CORS review indicated that
there is a misunderstanding on why the Vagrantfile in ironic exists.
In those comments, the assumption appeared to be that vagrant is used
to provision ironic, not just the development resources necessary to
work on ironic. This patch adds documentation to clarify that.

Change-Id: I7d71adaba402a62343d1e0399ee53410b76c973f