Updated openstack/openstack Project: openstack/swift  b32abab56c38ec72d290eda41eb19ca08d67c6bd
Allow default reseller prefix in domain_remap middleware

Previously, the reseller prefix needed to be provided in the host name
even when the domain was unique to that reseller. With the
default_reseller_prefix, any domain which matches in this middleware,
will will be passed on with a reseller prefix, whether or not it was
provided.

Change-Id: I5aa5ce78ad1ee2e3660cce4c3e07306f8999f02a
Implements: blueprint domainremap-reseller-domains