Updated openstack/openstack Project: openstack/glance_store  b529f0a782cc751a7ce35ebf9c2a929a8134aae6
Support V3 authentication with swift

This is the easiest way to support v3 authentication in the swift backend. It
allows us to specify that we want v3 authentication and provide the extra
authentication attributes to swiftclient.

It allows specifying auth_version via individual authentication references so
that you can mix v2 and v3 authentication.

Ideally in future we would move to a session orientated solution. In which case
we will need to handle the existing options in a backwards compatible way and
so I don't think that the extra arguments will cause any additional problems
there.

Change-Id: Ifb7c9be805689e938ff0e99a18fa220993862001