Updated openstack/openstack Project: openstack/openstack-doc-tools  072fc1d6190994b09dd7ef271e38c50b820e5fdd
Fix up README to work with the release tools

Add links to the bug tracker and source in a way the release script can
find them.

Move the contributing instructions to a separate file, where github will
pick them up and the release tools won't be confused by the urls.

Change-Id: Icbc1fdde3245ef6aa750bd44601adb1514763034