Updated openstack/openstack Project: openstack/ironic  a55f316df22d68cda884ad260231eef10c638b12
Change return value of [driver_]vendor_passthru to dict

This is a refactor patch that changes the return value of the
vendor_passthru() and driver_vendor_passthru() RPC methods to a dictionary
instead of a tuple. By doing that we make it easy to extend for future
improvements.

The RPC version was bumped to 1.29 by this change.

Change-Id: I8729bb85d97a563d4ec8409ecd0eba15dc0c51bc
Related-Blueprint: ipxe-dynamic-config