Updated openstack/openstack Project: openstack/ironic  95b71efa065bd98275f3b3f9715bfb1675b72ae9
Refactor node's and driver's vendor passthru to a common place

This patch is refactoring the node's and driver's vendor passthru in the
API to a common place (utils.py). Historically this code was duplicated
because each endpoint supported different things (driver's vendor
passhtru endpoing being always synchrounous and node's being always
asyncronous), since now both endpoints supports the same features and
the code is almost the same it makes sense to have it in a common place,
so changing once will affect both.

Related-Blueprint: ipxe-dynamic-config
Change-Id: I16be0c787993240e55ceb13b29bcb18c21151722