Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  e460a1e1ae58d68e764a7067ff1489dc18eb025b
Move writexml fix for older minidom to same module as caller

With the reorganization of code to move the XmlJob class to a separate
file, the patch to modify the minidom method writexml for earlier
instances that incorrectly added whitespace should be relocated to be
alongside.

This ensures the patch exists along with the code that is using
minidom's functionality.

Change-Id: I6ed3267c3c076dc5f8d40f0c97706c33e4018261