Updated openstack/openstack Project: openstack-dev/devstack  c342243cf803af7ab4c371eff3bff71ffb45d9ab
Add ability to specify cinder lvm_type option

Cinder has had an lvm_type option for quite a while
that allows the LVM driver to be configured to use
thin provisioning.

This required an additional PPA in Precise, but is
available in the default Trusty packages.

This patch adds the lvm_type option, and we'll use
it to do gate testing against the lvm_thin configuration.

Change-Id: I99c7fea131f3d79747ec75052adf8b24f41ba483